+++ tests/test_signoff_flow.py
@@ -1,3 +1,4 @@
+import datetime
 import random
 import string
 import unittest
@@ -285,27 +286,51 @@ class RefreshSignatureTest(SignoffWebTest, unittest.TestCase):
 
 class TrackingFieldsTest(SignoffWebTest, unittest.TestCase):
 
-    def last_author_is_tracked(self):
+    today = datetime.date.today().strftime("%Y-%m-%d")
+
+    def last_edit_by_and_date_are_tracked(self):
         self.app.post_json(self.source_collection + "/records",
                            {"data": {"title": "Hallo"}},
                            headers=self.headers)
         resp = self.app.get(self.source_collection, headers=self.headers)
-        assert resp.json["data"]["last_author"] == self.userid
+        assert resp.json["data"]["last_edit_by"] == self.userid
+        assert resp.json["data"]["last_edit_date"].startswith(self.today)
 
-    def test_last_editor_is_tracked(self):
+    def test_last_review_request_by_and_date_are_tracked(self):
         self.app.patch_json(self.source_collection,
                             {"data": {"status": "to-review"}},
                             headers=self.headers)
         resp = self.app.get(self.source_collection, headers=self.headers)
-        assert resp.json["data"]["last_editor"] == self.userid
+        assert resp.json["data"]["last_review_request_by"] == self.userid
+        assert resp.json["data"]["last_review_request_date"].startswith(self.today)
 
-    def test_last_reviewer_is_tracked(self):
+    def test_last_review_by_and_date_are_tracked(self):
         self.app.patch_json(self.source_collection,
                             {"data": {"status": "to-sign"}},
                             headers=self.headers)
         resp = self.app.get(self.source_collection, headers=self.headers)
         assert resp.json["data"]["status"] == "signed"
-        assert resp.json["data"]["last_reviewer"] == self.userid
+        assert resp.json["data"]["last_review_by"] == self.userid
+        assert resp.json["data"]["last_review_date"].startswith(self.today)
+        assert resp.json["data"]["last_signature_by"] == self.userid
+        assert resp.json["data"]["last_signature_date"].startswith(self.today)
+
+    def test_last_review_differs_from_last_signature_on_refresh_signature(self):
+        self.app.patch_json(self.source_collection,
+                            {"data": {"status": "to-sign"}},
+                            headers=self.headers)
+        resp = self.app.get(self.source_collection, headers=self.headers)
+        assert resp.json["data"]["status"] == "signed"
+        last_reviewer = resp.json["data"]["last_review_by"]
+
+        self.app.patch_json(self.source_collection,
+                            {"data": {"status": "to-sign"}},
+                            headers=self.headers)
+        metadata = self.app.get(self.source_collection, headers=self.headers).json["data"]
+        assert metadata["status"] == "signed"
+
+        assert metadata["last_signature_date"] != metadata["last_review_date"]
+        assert last_reviewer == metadata["last_review_by"]
 
     def test_editor_can_be_reviewer(self):
         self.app.patch_json(self.source_collection,
@@ -334,7 +359,9 @@ class TrackingFieldsTest(SignoffWebTest, unittest.TestCase):
         assert source_collection["status"] == "signed"
 
         # All tracking fields are here.
-        expected = ("last_author", "last_editor", "last_reviewer")
+        expected = ("last_edit_by", "last_edit_date", "last_review_request_by",
+                    "last_review_request_date", "last_review_by", "last_review_date",
+                    "last_signature_by", "last_signature_date")
         assert all([f in source_collection for f in expected])
 
         # They cannot be changed nor removed.